Added Windows Registry searcher
